+++ client/src/components/idea/discussion/CommentWriteBox.js
@@ -3,6 +3,7 @@ import axios from "axios";
 import {AppContext, BoardContext, IdeaContext} from "context";
 import React, {useContext, useState} from "react";
 import TextareaAutosize from "react-autosize-textarea";
+import {FaTimes} from "react-icons/fa";
 import tinycolor from "tinycolor2";
 import {UiClickableTip, UiPrettyUsername} from "ui";
 import {UiLoadableButton} from "ui/button";
@@ -10,6 +11,26 @@ import {UiFormControl, UiMarkdownFormControl} from "ui/form";
 import {UiCol} from "ui/grid";
 import {UiAvatar} from "ui/image";
 import {popupError, popupWarning} from "utils/basic-utils";
+
+const ReplyMarkdownOptions = styled.div`
+  border-bottom-left-radius: 0 !important;
+  border-bottom-right-radius: 0 !important;
+`;
+
+const ReplyBox = styled.div`
+  border-bottom-right-radius: var(--border-radius);
+  border-bottom-left-radius: var(--border-radius);
+  background-color: vaR(--tertiary);
+  font-size: small;
+  padding: .25rem .5rem;
+`;
+
+const ReplyCancelButton = styled(FaTimes)`
+  float: right;
+  margin-top: .25rem;
+  color: ${props => props.theme};
+  cursor: pointer;
+`;
 
 const WriteBox = styled(UiCol)`
   display: inline-flex;
@@ -23,8 +44,8 @@ const UsernameBox = styled.small`
   font-weight: bold;
 `;
 
-const CommentWriteBox = ({onCommentSubmit}) => {
-    const {user} = useContext(AppContext);
+const CommentWriteBox = ({onCommentSubmit, replyTo, setReplyTo}) => {
+    const {user, getTheme} = useContext(AppContext);
     const {onNotLoggedClick, data} = useContext(BoardContext);
     const {ideaData} = useContext(IdeaContext);
     const isModerator = data.moderators.find(mod => mod.userId === user.data.id);
@@ -44,6 +65,7 @@ const CommentWriteBox = ({onCommentSubmit}) => {
             ideaId: ideaData.id,
             description: message,
             type,
+            replyTo: replyTo.id
         }).then(res => {
             if (res.status !== 200 && res.status !== 201) {
                 popupError();
@@ -89,25 +111,37 @@ const CommentWriteBox = ({onCommentSubmit}) => {
     };
     const submitButton = user.loggedIn ? renderSubmitButton : () => void 0;
     const renderWriteCommentForm = () => {
-      if(ideaData.commentingRestricted && !isModerator) {
-          return <UiFormControl disabled className={"mt-1"} id={"commentMessage"} rows={1} maxRows={5} placeholder={"Commenting restricted to moderators only."}
-                                style={{overflow: "hidden"}} label={"Commenting restricted"}/>;
-      }
-      return <UiMarkdownFormControl as={TextareaAutosize} className={"mt-1"} id={"commentMessage"} rows={1} maxRows={5} placeholder={"Write a comment..."}
-                            style={{overflow: "hidden"}} onChange={onChange} label={"Write a comment"} onClick={onClick}/>;
-    };
-    return <WriteBox xs={10}>
-        <div className={"text-center mr-3 pt-2"}>
-            {avatar}
-            <br/>
+        if (ideaData.commentingRestricted && !isModerator) {
+            return <UiFormControl disabled className={"mt-1"} id={"commentMessage"} rows={1} maxRows={5} placeholder={"Commenting restricted to moderators only."}
+                                  style={{overflow: "hidden"}} label={"Commenting restricted"}/>;
+        }
+        return <div id={"replyBox"}>
+            <UiMarkdownFormControl CustomOptions={replyTo != null && ReplyMarkdownOptions} as={TextareaAutosize} className={"mt-1"} id={"commentMessage"} rows={1} maxRows={5} placeholder={"Write a comment..."}
+                                   style={{overflow: "hidden"}} onChange={onChange} label={"Write a comment"} onClick={onClick}/>
+            {replyTo != null && <React.Fragment>
+                <ReplyBox className={"text-black-60"}>
+                    <span className={"text-black-75"}>Reply to</span>
+                    <UiAvatar roundedCircle className={"ml-2 mr-1 move-top-1px"} size={16} user={replyTo.user} style={{minWidth: "16px"}}/>
+                    <strong><UiPrettyUsername user={replyTo.user}/></strong>
+                    <ReplyCancelButton theme={getTheme().toString()} onClick={() => setReplyTo(null)}/>
+                </ReplyBox>
+            </React.Fragment>}
         </div>
-        <UiCol xs={12} className={"px-0"}>
-            <UsernameBox>{username}</UsernameBox>
-            <br/>
-            {renderWriteCommentForm()}
-            {submitButton()}
-        </UiCol>
-    </WriteBox>
+    };
+    return <React.Fragment>
+        <WriteBox xs={10}>
+            <div className={"text-center mr-3 pt-2"}>
+                {avatar}
+                <br/>
+            </div>
+            <UiCol xs={12} className={"px-0"}>
+                <UsernameBox>{username}</UsernameBox>
+                <br/>
+                {renderWriteCommentForm()}
+                {submitButton()}
+            </UiCol>
+        </WriteBox>
+    </React.Fragment>
 };
 
 export default CommentWriteBox;